Fix docker-entrypoint.sh for HAProxy 1.8 upgrade.

+++ docker-entrypoint.sh
@@ -6,10 +6,18 @@ if [ "${1#-}" != "$1" ]; then
 fi
 
 if [ "$1" = 'haproxy' ]; then
-	# if the user wants "haproxy", let's use "haproxy-systemd-wrapper" instead so we can have proper reloadability implemented by upstream
-	shift # "haproxy"
-	rm -f /run/haproxy.pid
-	set -- "$(which haproxy-systemd-wrapper)" -p /run/haproxy.pid "$@"
+	if haproxy -v | grep -qF 'version 1.7.'; then
+		# if the user wants "haproxy", let's use "haproxy-systemd-wrapper" instead so we can have proper reloadability implemented by upstream
+		shift # "haproxy"
+		rm -f /run/haproxy.pid
+		set -- "$(which haproxy-systemd-wrapper)" -p /run/haproxy.pid "$@"
+	else:
+		shift # "haproxy"
+		# if the user wants "haproxy", let's add a couple useful flags
+		#   -W  -- "master-worker mode" (similar to the old "haproxy-systemd-wrapper"; allows for reload via "SIGUSR2")
+		#   -db -- disables background mode
+		set -- haproxy -W -db "$@"
+	fi
 fi
 
 PREFIX=docker-entrypoint.sh
